requirements.txt
appdirs==1.4.3
Django>=1.11,<2.0
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
six==1.10.0
-e git+https://github.com/praekelt/django-formfactory.git@develop#egg=django-formfactory
-e git+git@github.com:praekelt/django-export.git@600d061b9c893c12bb55962eed2f5323b2c7b77f#egg=django_export
django-object-tools==1.11.0
django-formtools==2.0
PyYAML==3.12
django-pagedown==0.1.3
django-markdown-deux==1.0.5
markdown2==2.3.4
django-crispy-forms==1.6.1
djangorestframework==3.6.4
